+++ packages/json-rpc-engine/src/createAsyncMiddleware.ts
@@ -40,6 +40,7 @@ type ReturnHandlerCallback = Parameters<JsonRpcEngineReturnHandler>[0];
  * The return handler will always be called. Its resolution of the promise
  * enables the control flow described above.
  *
+ * @deprecated Use `JsonRpcEngineV2` and its corresponding types instead.
  * @param asyncMiddleware - The asynchronous middleware function to wrap.
  * @returns The wrapped asynchronous middleware function, ready to be consumed
  * by JsonRpcEngine.